requirements.txt
torch>=2.2.0
torchvision>=0.17.0
timm>=0.9.12
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0
pillow>=10.2.0
matplotlib>=3.8.0
seaborn>=0.13.0
grad-cam>=1.5.0
shap>=0.45.0
flask>=3.0.0
kaggle>=1.6.0
albumentations>=1.4.0
opencv-python>=4.9.0
tqdm>=4.66.0
scipy>=1.12.0
